Add dexes to AAB (#19)
Adds precompiled dex files to the AAB by adding the directories containing the precompiled dex files to the input of the `mergeDexRelease` gradle task. The `mergeDexRelease` gradle task takes a file collection called `dexDirs` as input. The files in that file collection will all be merged together into a single dex file which is then packaged into the AAB.