Add `--xmanifest-path` argument to specify a separate manifest file (#28)
Allows a user of xbuild to specify a specific manifest.yaml to use when building. Useful for having a specific manifest for development and for release builds.

Ideally we shouldn't introduce this kind of duplication. A better solution would be to make xbuild more configurable through a build script. For example, when including vulkan validation layers for a development build, one should be able to use `xbuild:include-library=path`. Some more design work needs to go into this so for now we'll stick with the duplicate manifest yaml